+++ frontend/frontend_driver.c
@@ -309,7 +309,7 @@ enum frontend_architecture frontend_driver_get_cpu_architecture(void)
    return frontend->get_architecture();
 }
 
-uint32_t frontend_driver_get_total_memory(void)
+uint64_t frontend_driver_get_total_memory(void)
 {
    frontend_ctx_driver_t *frontend = frontend_get_ptr();
    if (!frontend || !frontend->get_total_mem)
@@ -317,7 +317,7 @@ uint32_t frontend_driver_get_total_memory(void)
    return frontend->get_total_mem();
 }
 
-uint32_t frontend_driver_get_free_memory(void)
+uint64_t frontend_driver_get_free_memory(void)
 {
    frontend_ctx_driver_t *frontend = frontend_get_ptr();
    if (!frontend || !frontend->get_free_mem)
