+++ frontend/frontend_driver.h
@@ -80,8 +80,8 @@ typedef struct frontend_ctx_driver
    enum frontend_architecture (*get_architecture)(void);
    enum frontend_powerstate (*get_powerstate)(int *seconds, int *percent);
    int  (*parse_drive_list)(void*);
-   uint32_t (*get_total_mem)(void);
-   uint32_t (*get_free_mem)(void);
+   uint64_t (*get_total_mem)(void);
+   uint64_t (*get_free_mem)(void);
 
    const char *ident;
 
@@ -155,9 +155,9 @@ bool frontend_driver_get_core_extension(char *s, size_t len);
 
 bool frontend_driver_get_salamander_basename(char *s, size_t len);
 
-uint32_t frontend_driver_get_total_memory(void);
+uint64_t frontend_driver_get_total_memory(void);
 
-uint32_t frontend_driver_get_free_memory(void);
+uint64_t frontend_driver_get_free_memory(void);
 
 RETRO_END_DECLS
